Fixed some more int conversion errors on windows.
MSVC is really picky about this :P

+++ src/game_state.h
@@ -108,7 +108,7 @@ private:
                     const EventData& event_data);
   void PopulateConditionInputs(ConditionInputs* condition_inputs,
                                const Character& character) const;
-  void PopulateCharacterAccessories(SceneDescription* scene, int renderable_id,
+  void PopulateCharacterAccessories(SceneDescription* scene, uint16_t renderable_id,
                                     const mathfu::mat4& character_matrix,
                                     int num_accessories, int damage,
                                     int health) const;
